PDF SUMMARIZER AI/requirements.txt
PyMuPDF==1.23.5
openai==1.14.3
python-dotenv==1.0.1
rich==13.7.1